view de new release melhorada. tentativa de atualizar saldo do caixa

+++ db/migrate/20130928132746_create_cashiers.rb
@@ -1,7 +1,7 @@
 class CreateCashiers < ActiveRecord::Migration
   def change
     create_table :cashiers do |t|
-      t.integer :opening_balance
+      t.decimal :opening_balance
       t.references :enterprise     
 
       t.timestamps
